move common components into their own folders
