+++ client/components/common/icon/icon.test.tsx
@@ -1,7 +1,7 @@
 import { render, screen } from '@testing-library/react';
 import { describe, expect, it } from 'vitest';
 
-import Icon from './icon';
+import Icon from '.';
 
 describe('Icon', () => {
 	it('renders the chevron svg', () => {
